Fix the A4 paper-size cast and hoist the TIFF read buffer
PageSize.Code is a UInt16Value, so the uint literal did not compile.

TiffFrameCounter allocated its read buffers with stackalloc inside the IFD
walk, which grows the stack once per frame; a TIFF with a long chain could
overflow it. One buffer now serves the whole walk, sliced per read.

+++ tools/FileListPageCounter/src/FileListPageCounter.Core/Reporting/WordReportBuilder.cs
@@ -283,7 +283,7 @@ public static class WordReportBuilder
             Width = (uint)PageWidthTwips,
             Height = (uint)PageHeightTwips,
             Orient = PageOrientationValues.Portrait,
-            Code = 9U // A4
+            Code = (ushort)9 // the Windows paper-size code for A4
         });
 
         section.AppendChild(new PageMargin
